[refactor] Create a file for the PassChecker in pylint.checker.base
